Refactor the script to support system wide installtions.
Instead of specifing the full path to each VM configuration, VMs are
now searched in a global directory $VMDIR specified im system.cfg. A
directory layout is enforced:

 vmdir/
  vm1/
   config.cfg
  vm2/
   config.cfg

Only directories containing a config.cfg are recognized. The $NAME of
a VM is now derived from the directory, e.g. 'vm1' or 'vm2'. It must
be removed from existing configurations.