Add support for the GOP <-> VNC bridge.
This is implemented as another boot method 'bootrom_gop'. The VNC port
is created as 59XX, where XX is replaced by the VMs ID. This is the same
way most wrapper scripts use. The 59XX port range should be more or less
save. The user can specify the resolution, if bhyve should block until a
VMC connection arrives and if an USB tablet should be emulated.