Rework the shutdown / reboot detection and handling.
Apparently there's a small delay between the moment a VM with a given
name is shutdown or even destroyed and the name can be used again. Work
around that by introducing a small artificial delay for now. An upstream
bug report will come in the next few days. While here simplify the logic
and make things a little bit more robust.