Rename the tap devices to the scheme tap$ID_$BRIDGE.
The bridge name is derived from the bridge naming scheme. Supported
schemes are:

* bridge_servicename
* bridge-servicname

All names not matching any of these schemes are taken as they are. That
may lead to problems with the 15 character limit of devices names.

Additionally a description what VM to what bridge the devices connects
is added to it.

This change makes the tap device creation more reliable, problems with
duplicated device names are now much more unlikely than before. And it
makes the `ifconfig` output of hosts with many VMs clearer.

This change is also e prerequisite for support several NICs per VM.