Add support for multiple NICs per VM.
A whitespace separated list of bridges can be given in the config file.
The script creates a NIC for every one of these and connects it to the
given bridge. The NICs are added in the same order as the bridges, e.g.
the first NIC connects to the first given bridge. The NICs are connected
to PCI slot 29 and below.

The implementation has some limitations, most to keep the logic simple.
Only one NIC can be added to a given bridge and there are no further
parameters which can be configured. Nevertheless there might be bugs.
If more flexibility is needed the script should be rewritten in more
suiteable language.