Save more state.
While we don't need this state inside the script, it's good to have for
debugging purposes and doesn't add next to no overhead. It also can be
used by external scripts.

While here move the state handling out of the reboot loop. There's no
need to update the state at every reboot, it's static across all runs
of the VM in the same session.