Implement a watchdog, restarting a VM if the guest crashed.
Add a virtio_console device to the VM. Inside the VM it shows up as a
TTY, the new script `startvm_watchdog_freebsd` listens on it and reacts
on events send by the host. Versions for other guest OS might be added
at a later time. In the host the console shows up as a Unix Domain
Socket. `startvm` is altered to send events to the guest and parse
them. Care is taken that the wtachdog can be shutdown gracefully.

The lifecycle is:

1. `startvm` starts the guest.
2. When the guest sends ARM over the console the watchdog is armed.
3. `startvm` sends ARMED to the guest.
4. Every 8 seconds `startvm` sends PING to guest.
5. The guest must anser with PONG within 8 seconds.
6. When 16 pings where missed, `startvm` tries to shutdown the guest.
7. If the guest doesn't shutdown within 120 seconds it's killed.
8. In both cases the VM is restarted.
9. If the guest want's to exit the watchdog it sends DISARM.
10. `startvm` anwers with DISARMED.