Add a simple post crash cleanup logic to the rc script.
If there's an id file in the runtime directory and there are no bhyve
processes running, we are most likely restarting after a hard crash.
Cleanup the runtime directory in that case to be sure that we are in
a clean state.

Note that this cleanup isn't strictly necessary, startvm will work
without it.